Getting ready for app group

+++ Marguerite/FileHelper.swift
@@ -15,12 +15,17 @@ class FileHelper: NSObject, SSZipArchiveDelegate {
     // The names of the files from the GTFS zip needed
     private static let gtfsFileNames = ["agency", "calendar_dates", "routes", "stop_times", "stops", "trips"]
     
-    // Path to latest zipped gtfs data
-    private static let tempZipPath = tempFolderPath.stringByAppendingPathComponent("temp.zip")
-    // Path to the folder where the downloaded gtfs zip is stored and unzipped
-    private static let tempFolderPath = (NSSearchPathForDirectoriesInDomains(.DocumentDirectory, .UserDomainMask, true)[0] as! String).stringByAppendingPathComponent("TempGTFSFiles")
-    
     private let fileManager = NSFileManager.defaultManager()
+    
+    // Path to latest zipped gtfs data
+    class func tempZipPath() -> String? {
+        return tempFolderPath()?.stringByAppendingPathComponent("temp.zip")
+    }
+    
+    // Path to the folder where the downloaded gtfs zip is stored and unzipped
+    class func tempFolderPath() -> String? {
+        return FileHelper.documentsPath()?.stringByAppendingString("TempGTFSFiles")
+    }
     
     /**
     Detects if the neccessary gtfs files are in place
@@ -39,8 +44,10 @@ class FileHelper: NSObject, SSZipArchiveDelegate {
             let task = NSURLSession.sharedSession().dataTaskWithURL(url, completionHandler: { (data: NSData?, response: NSURLResponse?, error: NSError?) -> Void in
                 if let data = data where error == nil {
                     println("Downloaded new gtfs data")
-                    data.writeToFile(FileHelper.tempZipPath, atomically: true)
-                    self.movedNewZipToTempFolder()
+                    if let path = FileHelper.tempZipPath() {
+                        data.writeToFile(path, atomically: true)
+                        self.movedNewZipToTempFolder()
+                    }
                 }
             })
             task.resume()
@@ -71,28 +78,29 @@ class FileHelper: NSObject, SSZipArchiveDelegate {
     Starts to unzip the gtfs zip file
     */
     private func movedNewZipToTempFolder() {
-        let tempFolder = FileHelper.tempFolderPath
-        SSZipArchive.unzipFileAtPath(FileHelper.tempZipPath, toDestination: tempFolder, progressHandler: nil) { (path: String?, succeeded: Bool, error: NSError?) -> Void in
-            var latestFileInTemp = self.getLatestFileInFolder(tempFolder)
-            var latestFileInCurrent = self.getLatestFileInFolder(Util.getTransitFilesBasepath())
-            println("--- New GTFS Data ---")
-            println(latestFileInTemp)
-            println(latestFileInCurrent)
-            println("Unzipped new data")
-            if latestFileInTemp.timeIntervalSinceDate(latestFileInCurrent) > 0 {
-                println("Replacing older data")
-                self.moveTempFilesToTransitFolder()
-                DefaultsHelper.keyIs(true, key: NeedsDatabaseUpdateKey)
-                DefaultsHelper.saveDataForKey(latestFileInTemp, key: "GTFS Date")
-            } else {
-                println("Keeping current data")
+        if let tempFolder = FileHelper.tempFolderPath(), let tempZipPath = FileHelper.tempZipPath() {
+            SSZipArchive.unzipFileAtPath(tempZipPath, toDestination: tempFolder, progressHandler: nil) { (path: String?, succeeded: Bool, error: NSError?) -> Void in
+                var latestFileInTemp = self.getLatestFileInFolder(tempFolder)
+                var latestFileInCurrent = self.getLatestFileInFolder(Util.getTransitFilesBasepath())
+                println("--- New GTFS Data ---")
+                println(latestFileInTemp)
+                println(latestFileInCurrent)
+                println("Unzipped new data")
+                if latestFileInTemp.timeIntervalSinceDate(latestFileInCurrent) > 0 {
+                    println("Replacing older data")
+                    self.moveTempFilesToTransitFolder()
+                    DefaultsHelper.keyIs(true, key: NeedsDatabaseUpdateKey)
+                    DefaultsHelper.saveDataForKey(latestFileInTemp, key: "GTFS Date")
+                } else {
+                    println("Keeping current data")
+                }
+                println("Clearing temp folder")
+                let contents = self.fileManager.contentsOfDirectoryAtPath(tempFolder, error: nil) as! [String]
+                for fileName in contents {
+                    self.fileManager.removeItemAtPath(tempFolder.stringByAppendingPathComponent(fileName), error: nil)
+                }
+                println("--- Handled GTFS Data ---")
             }
-            println("Clearing temp folder")
-            let contents = self.fileManager.contentsOfDirectoryAtPath(tempFolder, error: nil) as! [String]
-            for fileName in contents {
-                self.fileManager.removeItemAtPath(tempFolder.stringByAppendingPathComponent(fileName), error: nil)
-            }
-            println("--- Handled GTFS Data ---")
         }
     }
     
@@ -100,13 +108,14 @@ class FileHelper: NSObject, SSZipArchiveDelegate {
     Moves the newer files to the main transit files folder
     */
     private func moveTempFilesToTransitFolder() {
-        let tempFolder = FileHelper.tempFolderPath
-        for fileName in FileHelper.gtfsFileNames {
-            let destinationPath = Util.getTransitFilesBasepath().stringByAppendingPathComponent(fileName + ".txt")
-            if NSFileManager.defaultManager().fileExistsAtPath(destinationPath) {
-                NSFileManager.defaultManager().removeItemAtPath(destinationPath, error: nil)
+        if let tempFolder = FileHelper.tempFolderPath() {
+            for fileName in FileHelper.gtfsFileNames {
+                let destinationPath = Util.getTransitFilesBasepath().stringByAppendingPathComponent(fileName + ".txt")
+                if NSFileManager.defaultManager().fileExistsAtPath(destinationPath) {
+                    NSFileManager.defaultManager().removeItemAtPath(destinationPath, error: nil)
+                }
+                NSFileManager.defaultManager().moveItemAtPath(tempFolder.stringByAppendingPathComponent(fileName + ".txt"), toPath: destinationPath, error: nil)
             }
-            NSFileManager.defaultManager().moveItemAtPath(tempFolder.stringByAppendingPathComponent(fileName + ".txt"), toPath: destinationPath, error: nil)
         }
     }
     
@@ -114,13 +123,15 @@ class FileHelper: NSObject, SSZipArchiveDelegate {
     Moves the bundled gtfs data to the temp folder as if downloaded
     */
     class func moveBundleToTempFolder() {
-        for fileNames in gtfsFileNames {
-            if let path = NSBundle.mainBundle().pathForResource("BundledGTFS", ofType: "zip") {
-                var isDir : ObjCBool = false
-                NSFileManager.defaultManager().copyItemAtPath(path, toPath: FileHelper.tempFolderPath.stringByAppendingPathComponent("temp.zip"), error: nil)
+        if let tempFolderPath = tempFolderPath() {
+            for fileNames in gtfsFileNames {
+                if let path = NSBundle.mainBundle().pathForResource("BundledGTFS", ofType: "zip") {
+                    var isDir : ObjCBool = false
+                    NSFileManager.defaultManager().copyItemAtPath(path, toPath: tempFolderPath.stringByAppendingPathComponent("temp.zip"), error: nil)
+                }
             }
+            ShuttleSystem.sharedInstance.fileHelper.movedNewZipToTempFolder()
         }
-        ShuttleSystem.sharedInstance.fileHelper.movedNewZipToTempFolder()
     }
     
     /**
@@ -137,11 +148,25 @@ class FileHelper: NSObject, SSZipArchiveDelegate {
     If the folder exists, great do nothing. If it doesn't exist create it.
     */
     private class func checkFolder(folderName: String) {
-        let path = (NSSearchPathForDirectoriesInDomains(.DocumentDirectory, .UserDomainMask, true)[0] as! NSString).stringByAppendingPathComponent(folderName)
-        let fileManager = NSFileManager.defaultManager()
-        var isDir : ObjCBool = false
-        if !fileManager.fileExistsAtPath(path, isDirectory:&isDir) {
-            fileManager.createDirectoryAtPath(path, withIntermediateDirectories: true, attributes: nil, error: nil)
+        if let documentsPath = documentsPath() {
+            let path = (documentsPath as NSString).stringByAppendingPathComponent(folderName)
+            println(path)
+            let fileManager = NSFileManager.defaultManager()
+            var isDir : ObjCBool = false
+            if !fileManager.fileExistsAtPath(path, isDirectory:&isDir) {
+                var error: NSError?
+                fileManager.createDirectoryAtPath(path, withIntermediateDirectories: true, attributes: nil, error: &error)
+                println(error)
+            }
         }
+        
     }
+    
+    class func documentsPath() -> String? {
+        if let url = NSFileManager.defaultManager().containerURLForSecurityApplicationGroupIdentifier("group.edu.stanford.Marguerite") {
+            return url.absoluteString
+        }
+        return nil
+    }
+    
 }
